Refactor layout to Layout
