+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from 'client/components/layout';
+import Layout from 'client/components/Layout';
 import { AppProps } from 'next/app';
 
 const App = ({ Component, pageProps }: AppProps) => {
